Finish the backend routes (#4)
Add in the functionality for the backend routes, using the new Firestore database. In order to get this out faster, unit tests will come in another update. However, all routes have been manually tested, edge cases included.

Changes to the API:
- request ids are strings since that is what is generated by Firestore
- dates used in queries cannot be exact values, but ranges which is why there is now a start and end date within the date field
- primary keys for feedback and requests cannot be stored in the table upon adding since they are generated when the entry is added to the relevant table. These fields have been removed from the types.

NOTE: I have not tested LOCATION routes, since it is still extremely ambiguous as to what we will do with this route