Fix incorrect api results (#55)
Updates:

Only call PlaceSearch when the user hits enter
Currently the code would do the following:

Use PlaceSearch to get a nearby location based on its coordinates
Strip the reponse object of its coordinates and just retain its name
When the user clicks that name in the dropdown, try to figure out the coordinates again by calling GeoCoding API
(Which is just unnecessary extra work to get information we already had in the first place!
FIX: add coordinates (and address for nice ui) to the recent location, ride request, and accept request object. This will allows us to do one place search or distance api call and pass around those coordinates and address returned without having to reengineer it ourselves!
Move the place search calls to the server in hopes that it would help the app crashing (it did not). But it is still good practice to do those calls in the server!