Finish driver handle (#78)
In this branch we:

Added the styling for all four phases of the handle ride component on the driver's side
Added the proper callbacks given to us to implement the logic for switching between phases and dictating what UI is shown depending on the driver's location(cancel buttons, I've picked up student, etc) and talking back to the server so student knows where driver is.
Added the progress bar based on the example on the student side. Note: the progress bar has not been thoroughly tested, and I consider it to be sus code.