fix profile and legend. couldn't fix progress bar (#82)
Updates:
Profile now uses a View instead of an image circle to hold the person user
Legend width is based on the rendering size needed, not a fixed width
I cannot for the life of me fix the progress bar flex, so I have given up. This isn't a priority so we can do it later if needed.