Cancel ride bug (#114)
Fixes:
switched from using to component. This is because renders every item visually at once, creating a lot of overhead. We have 100+ locations. That's too much. is more dynamic, only VISUALLY rendering elements that are visible (plus some buffer), and caches the rest, displaying them only when needed.
Set a defined height for the container that contains all the list of locations. This is because since we have a dynamic number of elements being displayed, it's hard to predict how large the "scroll/load size" should be, making it difficult for both and to render its elements. Therefore, by providing a set height, it allows the components to know exactly how many elements to visually render.

SNIGDHA updates:
make it so that when driver cancels, the student gets notified
if the same notification is sent multiple times, the current code wouldn't show the subsequent notifications (since they have the same message). Change the trigger to re-show the notif be the timestamp instead of the text so we can retrigger notifications!
make sure driver location is not set to 0 on cancel (which made them not able to accept more rides)