Fixed View Ride race condition (#148)
There was an issue where two drivers could try to view one ride. Here is what was happening:

-> Both drivers requests see "Hey, some ride exists, I will go view it"
-> One of the driver's transactions wins and gets to view the ride, okay.
-> Other driver's transaction sees now the ride it wanted to view is taken (sad, error).

Okay. So now this case is handled by doing another thing:

-> Other driver's request will see the problem, and retry.
-> If retrying is going on too many times, it will give up and tell driver there is no rides.

Based on number of users, exceeding retry limit is probably indications of some other bug, so maybe we want some special behavior here, but the race condition is still fixed. In this case, driver can now just click the button again, and server will not kill itself with an unhandled error. Life is good.